revert uint48 change

+++ protocol/contracts/beanstalk/migration/L1AppStorage.sol
@@ -428,7 +428,7 @@ contract Storage {
     struct SiloSettings {
         bytes4 selector; // ────────────────────┐ 4
         uint32 stalkEarnedPerSeason; //         │ 4  (8)
-        uint48 stalkIssuedPerBdv; //            │ 4  (12)
+        uint32 stalkIssuedPerBdv; //            │ 4  (12)
         uint32 milestoneSeason; //              │ 4  (16)
         int96 milestoneStem; //                 │ 12 (28)
         bytes1 encodeType; //                   │ 1  (29)
